rename out_features to features for equivariant
