+++ netket/nn/symmetric_linear.py
@@ -331,7 +331,7 @@ class DenseEquivariantFFT(Module):
 
     product_table: HashableArray
     """ product table for space group"""
-    out_features: int
+    features: int
     """The number of output features. Will be the second dimension of the output."""
     shape: Tuple
     """Tuple that corresponds to shape of lattice"""
@@ -366,8 +366,8 @@ class DenseEquivariantFFT(Module):
         )
 
     def make_kernel(self, kernel):
-        """Converts the convolutional kernel of shape (out_features, in_features, n_symm)
-        to the expanded kernel of shape (out_features, in_features, n_point(in),
+        """Converts the convolutional kernel of shape (features, in_features, n_symm)
+        to the expanded kernel of shape (features, in_features, n_point(in),
         n_point(out), *shape) used in FFT-based group convolutions."""
         kernel = kernel[..., self.mapping]
 
@@ -391,7 +391,7 @@ class DenseEquivariantFFT(Module):
             "kernel",
             self.kernel_init,
             (
-                self.out_features,
+                self.features,
                 in_features,
                 self.n_point * self.n_cells,
             ),
@@ -422,7 +422,7 @@ class DenseEquivariantFFT(Module):
         x = x.reshape(*x.shape[:2], -1)
 
         if self.use_bias:
-            bias = self.param("bias", self.bias_init, (self.out_features,), self.dtype)
+            bias = self.param("bias", self.bias_init, (self.features,), self.dtype)
             bias = jnp.asarray(bias, dtype)
             x += jnp.expand_dims(bias, (0, 2))
 
@@ -437,7 +437,7 @@ class DenseEquivariantIrrep(Module):
     representations of the group.
 
     Acts on a feature map of shape [batch_size, in_features, n_symm] and
-    eeturns a feature map of shape [batch_size, out_features, n_symm].
+    eeturns a feature map of shape [batch_size, features, n_symm].
     The input and the output are related by
 
     .. math ::
@@ -462,7 +462,7 @@ class DenseEquivariantIrrep(Module):
     """Irrep matrices of the symmetry group. Each element of the list is an
     array of shape [n_symm, d, d]; irreps[i][j] is the representation of the
     jth group element in irrep #i."""
-    out_features: int
+    features: int
     """The number of output features. Will be the second dimension of the output."""
     use_bias: bool = True
     """Whether to add a bias to the output (default: True)."""
@@ -569,7 +569,7 @@ class DenseEquivariantIrrep(Module):
         kernel = self.param(
             "kernel",
             self.kernel_init,
-            (self.out_features, in_features, self.n_symm),
+            (self.features, in_features, self.n_symm),
             self.dtype,
         )
 
@@ -590,7 +590,7 @@ class DenseEquivariantIrrep(Module):
         x = self.inverse_ft(x)
 
         if self.use_bias:
-            bias = self.param("bias", self.bias_init, (self.out_features,), self.dtype)
+            bias = self.param("bias", self.bias_init, (self.features,), self.dtype)
             bias = jnp.asarray(bias, dtype)
 
             x += jnp.expand_dims(bias, (0, 2))
@@ -609,9 +609,9 @@ class DenseEquivariantMatrix(Module):
     """Flattened product table generated by PermutationGroup.produt_table().ravel()
     that specifies the product of the group with its involution, or the
     PermutationGroup object itself"""
-    out_features: int
+    features: int
     """The number of symmetry-reduced output features. The full output size
-    is n_symm*out_features."""
+    is n_symm*features."""
     use_bias: bool = True
     """Whether to add a bias to the output (default: True)."""
     mask: Optional[HashableArray] = None
@@ -632,15 +632,15 @@ class DenseEquivariantMatrix(Module):
 
     def full_kernel(self, kernel):
         """
-        Converts the convolutional kernel of shape (out_features, in_features, n_symm)
-        to a full dense kernel of shape (in_features*n_symm, out_features*n_symm).
+        Converts the convolutional kernel of shape (features, in_features, n_symm)
+        to a full dense kernel of shape (in_features*n_symm, features*n_symm).
         """
         in_features = kernel.shape[1]
         # result[out, in, g, h] == kernel[out, in, g^{-1}h]
         result = jnp.take(kernel, jnp.asarray(self.product_table), 2)
         # collate input dimensions [in, g] and output dimensions [out, h]
         result = result.transpose(1, 2, 0, 3).reshape(
-            self.n_symm * in_features, self.n_symm * self.out_features
+            self.n_symm * in_features, self.n_symm * self.features
         )
 
         return result
@@ -663,7 +663,7 @@ class DenseEquivariantMatrix(Module):
         kernel = self.param(
             "kernel",
             self.kernel_init,
-            (self.out_features, in_features, self.n_symm),
+            (self.features, in_features, self.n_symm),
             self.dtype,
         )
 
@@ -682,10 +682,10 @@ class DenseEquivariantMatrix(Module):
             precision=self.precision,
         )
 
-        x = x.reshape(-1, self.out_features, self.n_symm)
+        x = x.reshape(-1, self.features, self.n_symm)
 
         if self.use_bias:
-            bias = self.param("bias", self.bias_init, (self.out_features,), self.dtype)
+            bias = self.param("bias", self.bias_init, (self.features,), self.dtype)
             x += jnp.expand_dims(bias, (0, 2))
 
         return x
@@ -765,12 +765,18 @@ def DenseSymm(symmetries, point_group=None, mode="auto", shape=None, **kwargs):
 
 
 def DenseEquivariant(
-    symmetries, mode="auto", shape=None, point_group=None, in_features=None, **kwargs
+    symmetries,
+    features: int = None,
+    mode="auto",
+    shape=None,
+    point_group=None,
+    in_features=None,
+    **kwargs,
 ):
     r"""A group convolution operation that is equivariant over a symmetry group.
 
     Acts on a feature map of symmetry poses of shape [num_samples, in_features, num_symm]
-    and returns a feature  map of poses of shape [num_samples, out_features, num_symm]
+    and returns a feature  map of poses of shape [num_samples, features, num_symm]
 
     G-convolutions are described in ` Cohen et. {\it al} <http://proceedings.mlr.press/v48/cohenc16.pdf>`_
     and applied to quantum many-body problems in ` Roth et. {\it al} <https://arxiv.org/pdf/2104.05085.pdf>`_
@@ -784,6 +790,9 @@ def DenseEquivariant(
     Group elements that differ by the same symmetry operation (i.e. :math:`g = xh`
     and :math:`g' = xh'`) are connected by the same filter.
 
+    This layer maps an input of shape `(..., in_features, n_sites)` to an
+    output of shape `(..., features, num_symm)`.
+
     Args:
         symmetries: A specification of the symmetry group. Can be given by a
             nk.graph.Graph, an nk.utils.PermuationGroup, a list of irreducible
@@ -794,7 +803,7 @@ def DenseEquivariant(
             fourier transform over the translation group, a fourier transform using
             the irreducible representations or by constructing the full kernel matrix.
         shape: A tuple specifying the dimensions of the translation group.
-        out_features: The number of output features. The full output shape
+        features: The number of output features. The full output shape
             is [n_batch,features,n_symm].
         use_bias: A bool specifying whether to add a bias to the output (default: True).
         mask: An optional array of shape [n_sites] consisting of ones and zeros
@@ -809,10 +818,25 @@ def DenseEquivariant(
     if in_features is not None:
         warn_deprecation(
             (
-                "`in_features` is now automatically detected from the input."
+                "`in_features` is now automatically detected from the input and deprecated."
                 "Please remove it when calling `DenseEquivariant`."
             )
         )
+    if "out_features" in kwargs:
+        warn_deprecation(
+            "`out_features` has been renamed to `features` and the old name is "
+            "now deprecated. Please update your code."
+        )
+        if features is not None:
+            raise ValueError(
+                "You must only specify `features`. `out_features` is deprecated."
+            )
+        features = kwargs.pop("out_features")
+
+    if features is None:
+        raise ValueError("`features` not specified (the number of output features).")
+
+    kwargs["features"] = features
 
     if isinstance(symmetries, Lattice) and (
         point_group is not None or symmetries._point_group is not None
